Move protobuf classes into separate directory
