+++ serdes-native/src/main/java/io/ballerina/stdlib/serdes/Serializer.java
@@ -30,6 +30,7 @@ import io.ballerina.runtime.api.values.BMap;
 import io.ballerina.runtime.api.values.BObject;
 import io.ballerina.runtime.api.values.BString;
 import io.ballerina.runtime.api.values.BTypedesc;
+import io.ballerina.stdlib.serdes.protobuf.DataTypeMapper;
 
 import java.util.Locale;
 import java.util.Map;
