+++ serdes-native/src/main/java/io/ballerina/stdlib/serdes/protobuf/ProtobufMessage.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package io.ballerina.stdlib.serdes;
+package io.ballerina.stdlib.serdes.protobuf;
 
 import com.google.protobuf.DescriptorProtos;
